+++ src/Common/dcclite/Socket.h
@@ -201,8 +201,8 @@ namespace dcclite
 
 			bool IsOpen() const;
 
-			bool Send(const NetworkAddress &destination, const void *data, size_t size);
-			std::tuple<Status, size_t> Send(const void *data, size_t size);
+			bool Send(const NetworkAddress &destination, const void *data, size_t size) const;
+			std::tuple<Status, size_t> Send(const void *data, size_t size) const;
 
 			std::tuple<Status, int> Receive(NetworkAddress &sender, void *data, const int size, const bool truncate = false);
 			std::tuple<Status, int> Receive(void *data, int size);